Document#title and #execute_js first implementation.
Document#title returns the title of the document using the DOM.

Document#execute_js executes Javascript code in the context of the HTML document and
reparses the DOM/HTML afterwards. Thus, if the Javascript code made any changes to the
document, those changes will be reflected in the Nokogiri document.